Split role into staff-roles and end-user-roles using discriminator
