+++ data-access/src/infrastructure-services-impl/datastore/mongodb/models/member.ts
@@ -2,7 +2,7 @@ import { Schema, model, Model, ObjectId, PopulatedDoc, Types } from 'mongoose';
 import { Base, SubdocumentBase, NestedPath, Patterns } from '../../../../../seedwork/services-seedwork-datastore-mongodb/interfaces/base';
 import * as User from './user';
 import * as Community from './community';
-import * as Role from './role';
+import * as EndUserRole from './roles/end-user-role';
 
 export interface Account extends SubdocumentBase {
   firstName: string;
@@ -67,7 +67,7 @@ export interface Member extends Base {
   accounts: Types.DocumentArray<Account>;
   cybersourceCustomerId: string;
   customViews: Types.DocumentArray<CustomView>;
-  role?: PopulatedDoc<Role.Role> | ObjectId;
+  role?: PopulatedDoc<EndUserRole.EndUserRole> | ObjectId;
   profile: Profile;
 }
 
@@ -80,7 +80,7 @@ const schema = new Schema<Member, Model<Member>, Member>(
     },
     memberName: { type: String, required: true, maxlength: 200 },
     community: { type: Schema.Types.ObjectId, ref: Community.CommunityModel.modelName, required: true, index: true },
-    role: { type: Schema.Types.ObjectId, ref: Role.RoleModel.modelName, required: false, index: true },
+    role: { type: Schema.Types.ObjectId, ref: EndUserRole.EndUserRoleModel.modelName, required: false, index: true },
     accounts: { type: [AccountSchema], required: false },
     customViews: { type: [CustomViewSchema], required: false },
     cybersourceCustomerId: { type: String, required: false, maxlength: 50 },
